begin dismantling classes
